feat(content-sidebar): Add annotations activity type
* split out menu into sub-component
* write tests for new component
